First interesting program is 221013.
Problem with this program is that it starts a virgin node with a config, and thinks that config will be achieved. But it won't because it's waiting for another node also in the config. So it just waits. This then means that in the program generation I have to either constrain things even further or model whether or not the cluster can be achieved. *groan*.
Or I could just treat the {transitioner,_} as the same as ready, provided it stays stable for a while. That would certainly be the simplest.