Invert behaviour: do not leave mnesia running when we boot rabbit.
It turns out that if rabbit decides it needs to do some upgrades to mnesia then in asserts (but incorrectly calls it "ensure") mnesia is not running before it starts the upgrades (as part of taking a backup of the mnesia dir). Thus we really must make sure we do not start mnesia in clusterer, or, if we need to, we must also stop it before booting rabbit itself.