Detect when rabbit:boot|start fails, catch, and report back to the coordinator who tries (10 seconds later) to boot again.
This is primarily expected to be a means to get around issues with upgrading clusters where some nodes come back online having been upgraded but now find their version differs from some of the older nodes who are still running. Eventually, the older nodes will also be upgraded and when the last to be upgraded gets upgraded and restarted, it will be the last man standing and should come all the way back online, allowing the others then to start.