Move from `.app.src` to `PROJECT_*` Makefile variables
This is recommended by Erlang.mk user guide and we can do it because
there is no application env variables for this plugin.

This allows to specify the plugin version from the command line:
    make PROJECT_VERSION=1.0.0

This was already done for the `stable` branch in commit b62698af47 but it
was not possible in the `master` branch yet because Erlang.mk didn't
allow to set `env` application variables or non-standard keys in the
`.app` file. Now that this is supported, we can move the `master` branch
to that new mechanism too.

[#130992027]